requirements.txt
asgiref==3.5.2
attrs==21.4.0
Django==4.0.5
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
environs==9.5.0
inflection==0.5.1
jsonschema==4.6.1
marshmallow==3.17.0
packaging==21.3
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
whitenoise==6.2.0
